Share backend and dependencies
