+++ app/ide-desktop/lib/common/src/text/english.json
