+++ app/ide-desktop/lib/dashboard/src/components/MenuEntry.tsx
@@ -4,7 +4,7 @@ import * as React from 'react'
 import BlankIcon from 'enso-assets/blank.svg'
 import * as detect from 'enso-common/src/detect'
 
-import type * as text from '#/text'
+import type * as text from 'enso-common/src/text'
 
 import type * as inputBindings from '#/configurations/inputBindings'
