+++ app/ide-desktop/lib/dashboard/src/hooks/toastAndLogHooks.ts
@@ -3,7 +3,7 @@ import * as React from 'react'
 
 import * as toastify from 'react-toastify'
 
-import type * as text from '#/text'
+import type * as text from 'enso-common/src/text'
 
 import * as loggerProvider from '#/providers/LoggerProvider'
 import * as textProvider from '#/providers/TextProvider'
