+++ app/ide-desktop/lib/dashboard/src/layouts/TabBar.tsx
@@ -4,7 +4,7 @@ import * as React from 'react'
 import * as reactQuery from '@tanstack/react-query'
 import invariant from 'tiny-invariant'
 
-import type * as text from '#/text'
+import type * as text from 'enso-common/src/text'
 
 import * as textProvider from '#/providers/TextProvider'
