+++ app/ide-desktop/lib/dashboard/src/modals/AboutModal.tsx
@@ -3,7 +3,7 @@ import * as React from 'react'
 
 import LogoIcon from 'enso-assets/enso_logo.svg'
 
-import type * as text from '#/text'
+import type * as text from 'enso-common/src/text'
 
 import * as backendProvider from '#/providers/BackendProvider'
 import * as textProvider from '#/providers/TextProvider'
