+++ app/ide-desktop/lib/dashboard/src/pages/subscribe/Subscribe/components/Card.tsx
@@ -7,7 +7,7 @@ import * as React from 'react'
 
 import Check from 'enso-assets/check_mark.svg'
 
-import type * as text from '#/text'
+import type * as text from 'enso-common/src/text'
 
 import * as textProvider from '#/providers/TextProvider'
